requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
mysqlclient>=2.1,<3.0
djangorestframework-simplejwt>=5.2,<6.0
redis>=5.0,<6.0



